Fix: addProductViewCount메소드 수정
redis에서 가져온 value값의 데이터 타입이 string 인데 Long 으로 처리하려고 해서 에러 발생, 형변환으로 수정함

+++ src/main/java/org/example/sivillage/product/domain/ProductViewCount.java
@@ -16,13 +16,13 @@ public class ProductViewCount {
     private Long id;
 
     @Column(nullable = false)
-    private Double viewCount;
+    private Long viewCount;
 
     @Column(nullable = false)
     private String productCode;
 
     @Builder
-    public ProductViewCount(Long id, Double viewCount, String productCode) {
+    public ProductViewCount(Long id, Long viewCount, String productCode) {
         this.id = id;
         this.viewCount = viewCount;
         this.productCode = productCode;
